backend/requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.22.0
pydantic==2.5.2
httpx==0.24.1
python-multipart==0.0.6
python-dotenv==1.0.0

